Fix `update_dashboard` #76: Correct Payload (#77)
* Fix `update_dashboard` #76: Correct Payload

Turns out during my testing, I forgot to have a permutation where I
updated the permissions of a dashboard. Discovered the payload is
malformed because of a lack of post-processing in the `jira` lib (this
is as intended). This PR fixes the issue.

* Update keys in payload dict for `update_dashboard` action from
`edit_permissions/share_permissions` to
`editPermissions/sharePermissions` to make the `jira` API happy.

* Force Tests

* Force Tests